requirements.txt
streamlit>=1.30.0
pyarrow>=10.0.0
pandas>=2.0.0
plotly>=5.15.0
scipy>=1.10.0
Pillow>=9.0.0
numpy>=1.24.0
